LayoutTests/http/wpt/web-animations/interfaces/Animatable/animate-no-browsing-context-expected.txt


PASS Element.animate() creates an animation with the correct timeline when called on an element in a document without a browsing context 
PASS The timeline associated with an animation trigger on an element in a document without a browsing context is inactive 
FAIL Replacing the timeline of an animation targetting an element in a document without a browsing context leaves it in the pending state assert_equals: The animation should be initially pending expected "pending" but got "finished"
FAIL Replacing the timeline of an animation targetting an element in a document without a browsing context and then adopting that element causes it to start updating style assert_equals: Style should be updated expected "0" but got "0.3499999940395355"
